test: keep reader metrics in one class
